Move Git integration tests to the test/integration pkg
